Big refactoring started
