--- components/HeaderOld.tsx
@@ -1,39 +1,0 @@
-import Link from "next/link";
-import NavMenu, { Items } from "./navMenu/NavMenu";
-
-type Props = {
-  navList: Items[];
-};
-
-const Header: React.FC<Props> = ({ navList }: Props) => {
-  return (
-    <header className="w-full">
-      <div className="container mx-auto max-w-[1360px] px-5 py-[20px]">
-        <div className="flex gap-[48px] items-center">
-          <Link href="/" className="flex-none">
-            <div>
-              <svg
-                width="80"
-                height="60"
-                viewBox="0 0 80 60"
-                fill="none"
-                xmlns="http://www.w3.org/2000/svg"
-              >
-                <path
-                  d="M0.251999 3.98948C0.169662 3.20876 0.140105 0.619426 0.208789 0.203918C0.228097 0.0871614 0.248979 -0.00162218 0.255188 0.00662046C0.261397 0.0148631 0.496243 0.287966 0.777062 0.613507C3.63004 3.92069 7.77182 6.67251 12.0818 8.12448C13.9067 8.73926 15.2286 8.9936 17.0152 9.07379C18.2175 9.12772 18.8662 9.2468 19.8189 9.5883C21.1663 10.0713 22.6649 10.9971 23.7027 11.9877C23.8435 12.1221 23.9769 12.2321 23.999 12.2321C24.0211 12.2321 24.0856 12.1297 24.1422 12.0045C24.293 11.6712 24.8382 10.829 25.2331 10.3194C25.6475 9.78454 26.5017 8.82862 26.8001 8.56577L27.016 8.37567L26.9595 8.73646C26.1461 13.937 26.6116 19.1107 28.286 23.4781C28.4339 23.8639 28.6589 24.4076 28.7859 24.6863C29.0105 25.1789 29.4713 26.0707 29.5401 26.1461C29.5589 26.1667 29.6799 26.103 29.8091 26.0047C30.1361 25.7557 30.6837 25.418 31.4513 24.9921C32.1139 24.6245 32.471 24.4447 33.6697 23.8756L34.3333 23.5606L34.0853 23.2223C33.6254 22.5947 32.8677 21.164 32.2822 19.8171C32.0117 19.195 31.9946 19.1467 32.06 19.1878C32.0892 19.2061 32.3953 19.4038 32.7403 19.6271C33.0853 19.8505 33.3674 20.029 33.3671 20.0238C33.3668 20.0186 33.3281 19.888 33.2811 19.7335C33.1265 19.225 33.0628 18.6861 33.0394 17.6901C33.0168 16.7232 33.0567 15.9053 33.1794 14.8189C33.2657 14.0545 33.4582 12.8459 33.4936 12.8459C33.5098 12.8459 33.5466 13.0042 33.5753 13.1977C33.6465 13.6777 33.8411 14.3745 34.075 14.9874C34.4396 15.9429 35.1766 17.4117 35.782 18.3896L35.9976 18.7378L36.1634 18.3677C36.2545 18.1642 36.4995 17.6159 36.7076 17.1494C36.9936 16.5087 37.0942 16.321 37.119 16.3822C37.1461 16.449 37.8332 19.4325 37.8332 19.4833C37.8332 19.4921 37.9623 19.451 38.1202 19.3919C38.815 19.1319 39.9101 18.9858 40.5706 19.065C41.0656 19.1243 41.7063 19.2898 41.9628 19.4248C42.0585 19.4751 42.1422 19.51 42.1488 19.5023C42.1554 19.4946 42.319 18.7879 42.5125 17.9318C42.7059 17.0758 42.8761 16.3612 42.8907 16.3438C42.9052 16.3264 43.155 16.8456 43.4457 17.4975C43.7364 18.1494 43.9864 18.6905 44.0012 18.6999C44.0563 18.735 44.8739 17.2626 45.3407 16.2877C45.9836 14.9451 46.2523 14.179 46.4202 13.2101C46.5049 12.7212 46.5115 12.7296 46.6338 13.4817C47.0469 16.0219 47.083 18.4997 46.7244 19.6878C46.678 19.8412 46.6479 19.9748 46.6573 19.9845C46.6668 19.9942 46.9483 19.826 47.2829 19.6107C47.6175 19.3954 47.9146 19.205 47.9429 19.1877C48.0044 19.1502 47.9838 19.2036 47.6169 20.0356C47.0359 21.3526 46.3718 22.5879 45.8891 23.2493C45.7639 23.4209 45.6655 23.5643 45.6706 23.5677C45.6757 23.5712 45.9477 23.6983 46.2752 23.8502C47.7318 24.5257 49.2798 25.3685 50.0338 25.8966C50.4038 26.1558 50.4368 26.1703 50.4904 26.0985C50.6343 25.9055 51.4543 24.1551 51.6942 23.5288C53.1671 19.6829 53.7281 15.1888 53.2905 10.7414C53.2239 10.0644 53.0288 8.58357 52.9878 8.44305C52.9227 8.22041 54.1594 9.53069 54.8053 10.3688C55.1822 10.8578 55.5922 11.4916 55.8137 11.9273L55.9819 12.2582L56.4447 11.8499C57.7498 10.6983 59.1619 9.88902 60.6703 9.42822C61.3981 9.20585 61.9664 9.11891 62.9891 9.07353C64.0914 9.0246 64.6855 8.95208 65.6684 8.74641C69.2471 7.9976 73.3835 5.91335 76.4921 3.29258C77.5557 2.39589 78.655 1.30554 79.4266 0.381836L79.7456 0L79.7778 0.142361C79.8424 0.427653 79.8643 2.25151 79.8128 3.06876C79.5111 7.86747 77.9334 11.8374 75.2138 14.641C74.7064 15.1641 73.8308 15.9318 73.4283 16.2064C73.3347 16.2702 73.2645 16.329 73.2723 16.337C73.2964 16.3619 74.3516 16.0062 75.1218 15.7137C76.8301 15.0649 78.3054 14.3189 79.1641 13.6697C79.2968 13.5694 79.5345 13.3595 79.6924 13.2032C79.8504 13.047 79.9881 12.9279 79.9986 12.9387C80.009 12.9495 79.9618 13.1984 79.8937 13.4919C79.3655 15.7686 78.3697 18.4371 77.3411 20.3319C75.896 22.9943 74.2366 24.8651 72.2884 26.0286L71.9627 26.2231L72.303 26.1953C73.732 26.079 75.0595 25.7331 75.9343 25.2493C76.1071 25.1538 76.2581 25.0856 76.27 25.0979C76.3335 25.1634 75.8184 26.3683 75.3871 27.1628C73.4713 30.692 70.3217 33.4793 66.4191 35.0995C65.9306 35.3023 65.9331 35.3105 66.4978 35.3624C67.3532 35.4412 68.6772 35.1223 69.9585 34.5291C70.0959 34.4655 70.2184 34.4237 70.2307 34.4364C70.275 34.4821 70.0781 34.9819 69.9165 35.2336C68.9624 36.7201 66.4283 38.2167 62.7127 39.4878L61.7132 39.8298L62.0109 39.8668C62.5368 39.9321 64.3221 39.8742 64.9667 39.7708C65.2825 39.7202 65.5504 39.679 65.5621 39.6793C65.5738 39.6796 65.3109 39.8268 64.9778 40.0064C62.532 41.3251 59.8902 43.3848 57.8006 45.6023C54.643 48.9531 52.6232 52.896 52.0762 56.7775C51.9511 57.6647 51.9664 57.6059 51.8679 57.5794C51.7934 57.5593 51.7827 57.5198 51.7827 57.2645C51.7827 56.0145 51.2178 54.0602 50.3772 52.4026C49.9875 51.6341 49.6058 50.9804 49.5699 51.02C49.5239 51.0707 49.2614 51.9271 49.1088 52.5246C48.9488 53.1508 48.8482 53.7025 48.8482 53.9534C48.8482 54.2569 49.1371 54.9665 49.473 55.4881L49.6222 55.7197L49.5752 56.3472C49.4877 57.5154 49.4946 57.479 49.3602 57.479C49.2464 57.479 49.241 57.4897 49.1911 57.8188C49.1065 58.3767 48.9807 58.9736 48.8848 59.2719L48.7947 59.5522L48.5665 59.0579C48.2954 58.4706 47.939 57.9042 47.7014 57.6829C47.607 57.595 47.5298 57.5137 47.5298 57.5024C47.5298 57.4911 47.5775 57.456 47.6359 57.4245L47.7419 57.3672L47.3202 56.9341C46.9 56.5025 46.898 56.5011 46.7687 56.5564C46.6973 56.5869 46.5635 56.6961 46.4713 56.7992C46.3298 56.9575 46.2993 57.0263 46.2746 57.2437C46.1935 57.9558 45.9182 59.0116 45.6964 59.4609C45.6523 59.5503 45.5361 59.7081 45.4381 59.8117L45.26 60L45.2282 59.8027C45.2107 59.6942 45.1656 59.3628 45.1279 59.0662C44.9904 57.9851 44.7985 57.4307 44.288 56.64C44.0556 56.2802 43.7919 55.9883 43.699 55.9883C43.6755 55.9883 43.5805 56.0281 43.4878 56.0769C43.3287 56.1606 43.3194 56.1764 43.3194 56.3664C43.3194 56.5564 43.3037 56.5833 43.0311 56.8616C42.5582 57.3444 42.3048 57.9587 42.2973 58.6408L42.2944 58.9039L42.1082 58.5274C41.741 57.7849 41.4907 56.8648 41.4907 56.2574C41.4907 56.0657 41.4473 55.8098 41.3639 55.5088C41.2941 55.2572 41.2415 55.0488 41.2469 55.0458C41.4115 54.9523 41.9035 54.5519 42.1635 54.2997C42.5819 53.8939 42.8835 53.4821 43.0891 53.0358L43.2442 52.6991L42.8396 52.1844C42.6171 51.9013 42.39 51.6281 42.335 51.5774L42.2349 51.4851L41.0016 52.0047L39.7683 52.5243L38.5455 52.0202C37.8729 51.743 37.296 51.5162 37.2635 51.5162C37.2075 51.5162 36.8715 51.8839 36.411 52.4493L36.2045 52.7027L36.3827 53.0735C36.6843 53.701 37.2577 54.3605 37.9182 54.8396C38.0703 54.9499 38.1991 55.0427 38.2045 55.0458C38.21 55.0488 38.1573 55.2572 38.0876 55.5088C37.9982 55.831 37.9606 56.0634 37.96 56.2951C37.9587 56.8937 37.7542 57.6544 37.3873 58.4254L37.1596 58.9039L37.1555 58.6189C37.146 57.9538 36.8913 57.3441 36.4202 56.8584C36.1494 56.5792 36.132 56.5493 36.132 56.3634C36.132 56.1769 36.1222 56.1603 35.9636 56.0769C35.871 56.0281 35.7759 55.9883 35.7524 55.9883C35.6596 55.9883 35.3958 56.2802 35.1635 56.64C34.6486 57.4374 34.4542 58.0045 34.3219 59.0939C34.2848 59.3993 34.2415 59.7262 34.2257 59.8203L34.1969 59.9914L34.0694 59.8796C33.6998 59.5561 33.3339 58.4596 33.1752 57.2003C33.1544 57.0347 33.1111 56.9457 32.9805 56.7996C32.8882 56.6963 32.7542 56.5869 32.6828 56.5564C32.5534 56.5011 32.5514 56.5025 32.1312 56.9341L31.7095 57.3672L31.8156 57.4245C31.8739 57.456 31.9216 57.4911 31.9216 57.5024C31.9216 57.5137 31.8444 57.595 31.75 57.6829C31.5103 57.9061 31.1534 58.4752 30.8842 59.0632L30.6561 59.5616L30.5639 59.2547C30.4728 58.9508 30.2963 58.0891 30.2429 57.6872C30.2173 57.4944 30.2064 57.479 30.0945 57.479C29.9562 57.479 29.9641 57.5196 29.8762 56.3472L29.8293 55.7197L29.979 55.4871C30.1745 55.1837 30.4441 54.6248 30.5389 54.3271C30.6024 54.1273 30.6107 53.9822 30.5927 53.3845C30.5719 52.6898 30.4111 51.2221 30.3482 51.1511C30.2813 51.0758 29.4621 52.6569 29.08 53.5988C28.5456 54.916 28.2225 56.2717 28.2219 57.1998C28.2217 57.5194 28.2128 57.5588 28.1366 57.5793C28.0277 57.6087 28.0251 57.6027 27.9894 57.2378C27.9037 56.3615 27.582 54.9258 27.2232 53.818C25.8749 49.6558 22.9203 45.7037 18.7376 42.4679C17.443 41.4663 16.2011 40.6445 14.991 39.9887C14.6774 39.8188 14.4305 39.6796 14.4422 39.6793C14.4539 39.679 14.7218 39.7202 15.0376 39.7708C15.6871 39.8749 17.4752 39.9322 17.9934 39.8654L18.2911 39.827L17.6531 39.617C15.0636 38.7645 12.9967 37.7811 11.5313 36.7042C11.0536 36.3531 10.3511 35.6441 10.1047 35.2644C9.91115 34.9661 9.71948 34.4923 9.77085 34.4393C9.78302 34.4268 9.99217 34.5067 10.2357 34.6169C11.3522 35.1221 12.2915 35.3546 13.2089 35.3527C13.5246 35.3521 13.8267 35.333 13.8802 35.3103C13.9653 35.2741 13.9203 35.2454 13.5187 35.079C13.2664 34.9745 12.7538 34.7373 12.3795 34.5518C9.4216 33.0862 6.96802 30.9085 5.23798 28.2132C4.58252 27.192 4.09565 26.2348 3.8153 25.4159L3.68511 25.0356L4.03462 25.2301C4.93538 25.7314 6.23554 26.0745 7.70135 26.1977L8.04158 26.2262L7.65882 25.998C5.58008 24.7585 3.7239 22.5439 2.26018 19.5567C1.30268 17.6027 0.470938 15.2024 0.00596793 13.0512C-0.0180185 12.9403 0.0175358 12.9622 0.35292 13.2648C0.55842 13.4502 0.92753 13.739 1.17314 13.9066C1.76293 14.309 3.17332 15.0313 4.13856 15.4253C4.88401 15.7296 6.69579 16.3746 6.72352 16.3456C6.73092 16.3378 6.50419 16.1511 6.21972 15.9306C2.8484 13.3171 0.804408 9.22716 0.251999 3.98948Z"
-                  fill="#6C48D1"
-                />
-              </svg>
-            </div>
-          </Link>
-          <NavMenu
-            items={navList}
-            className={"flex font-batman gap-[40px]"}
-          ></NavMenu>
-        </div>
-      </div>
-    </header>
-  );
-};
-
-export default Header;
