--- components/header/burger.tsx
@@ -1,25 +1,0 @@
-const Burger = () => (
-  <div className="">
-    <svg
-      width="31"
-      height="17"
-      viewBox="0 0 31 17"
-      fill="none"
-      xmlns="http://www.w3.org/2000/svg"
-      data-v-09d20c72=""
-    >
-      <path
-        d="M0 11L23 11L18.0932 17L0 17V11Z"
-        fill="white"
-        data-v-09d20c72=""
-      ></path>
-      <path
-        d="M0 0L31 0L26.0179 6L0 6V0Z"
-        fill="white"
-        data-v-09d20c72=""
-      ></path>
-    </svg>
-  </div>
-);
-
-export default Burger;
