+++ src/components/Footer.tsx
@@ -1,4 +1,4 @@
-import NavMenu, { Items } from "./navMenu/NavMenu";
+// import NavMenu, { Items } from "./navMenu/NavMenu";
 interface Props {
   navList: Items[];
 }
@@ -8,7 +8,7 @@ const Footer: React.FC<Props> = ({ navList }) => {
     <footer className="w-full bg-main-bg">
       <div className="container mx-auto max-w-[1360px] px-5">
         <h2 className="text-3xl text-center">Footer</h2>
-        <NavMenu items={navList} className={"gap-[16px] font-batman"}></NavMenu>
+        {/* <NavMenu items={navList} className={"gap-[16px] font-batman"}></NavMenu> */}
       </div>
     </footer>
   );
