+++ src/components/icons-svg/IconArrowRight.tsx
