+++ src/components/icons-svg/IconVk.tsx
