+++ src/pages/_app.tsx
@@ -1,4 +1,5 @@
 import "../styles/globals.css";
+import "../styles/my-swiper.css";
 import type { AppProps } from "next/app";
 
 function MyApp({ Component, pageProps }: AppProps) {
