+++ src/styles/globals.css
@@ -7,9 +7,7 @@
   src: url("../fonts/Batman/batmfa__.woff");
 }
 
-@import "./_vars.css";
-@import "./my-swiper.css";
-
+/* ???? Определить что именно нужно */
 html,
 body {
   padding: 0;
@@ -26,35 +24,6 @@ a {
 
 * {
   box-sizing: border-box;
-}
-main {
-  z-index: 1;
-  position: relative;
-  margin-bottom: 372px;
-  margin-top: 112px;
-}
-
-header {
-  z-index: 999;
-  position: fixed;
-  top: 0;
-  background-color: #1c1c1e;
-}
-
-footer {
-  position: fixed;
-  bottom: 0;
-  width: 100%;
-  height: 372px;
-  background-color: #1c1c1e;
-}
-
-.font-batman {
-  font-family: "Batman Forever";
-}
-
-.header-title-section {
-  font-size: 50px;
 }
 
 @media (prefers-color-scheme: dark) {
