+++ src/styles/my-swiper.css
@@ -6,10 +6,10 @@
 .swiper-slide {
 }
 .swiper-slide img {
-  display: block;
+  /* display: block;
   width: 100%;
   height: 100%;
-  object-fit: cover;
+  object-fit: cover; */
 }
 
 .swiper-button-prev,
