+++ src/testData/HomePage/HomePage.ts
@@ -1,0 +1,45 @@
+import slideImg from "../../assets/testData/caruselExmpl.png";
+
+import { SlideData } from "../../types/globalTypes";
+export const SlidesData: SlideData[] = [
+  {
+    id: "1",
+    title: "Становись частью команды",
+    subTitle: "Вступай в новую команду киберспорта и развивайся в любимом деле",
+    img: slideImg,
+  },
+  {
+    id: "2",
+    title: "Становись частью команды",
+    subTitle: "Вступай в новую команду киберспорта и развивайся в любимом деле",
+    img: slideImg,
+  },
+  {
+    id: "3",
+    title: "Становись частью команды",
+    subTitle: "Вступай в новую команду киберспорта и развивайся в любимом деле",
+    img: slideImg,
+  },
+];
+
+import { ShortNewsCard } from "../../types/globalTypes";
+export const NewsData: ShortNewsCard[] = [
+  {
+    id: "1",
+    title: "Ребрендинг",
+    date: "Сентябрь 2022",
+    img: "",
+  },
+  {
+    id: "2",
+    title: "набор игроков",
+    date: "Октябрь 2022",
+    img: "",
+  },
+  {
+    id: "3",
+    title: "новый стаф",
+    date: "Октябрь 2022",
+    img: "",
+  },
+];
