+++ src/utils/headerData.ts
@@ -1,0 +1,32 @@
+import { navLink } from "../types/globalTypes";
+
+export const navLinks: navLink[] = [
+  {
+    name: "главная",
+    path: "/",
+  },
+  {
+    name: "о нас",
+    path: "/about",
+  },
+  {
+    name: "ивенты",
+    path: "/ivents",
+  },
+  {
+    name: "команда",
+    path: "/teams",
+  },
+  {
+    name: "достижения",
+    path: "/achivments",
+  },
+  {
+    name: "сотрудничество",
+    path: "/contact",
+  },
+  {
+    name: "twitch",
+    path: "/stream",
+  },
+];
